[Command-buffer][L0] Reset sync-point events
The L0 events used to implement UR sync-points in
a command-buffer aren't reset after the first execution.
Leaving them in a signalled state before subsequent submissions
of the UR command-buffer.

This patch resets the L0 events used as UR command-buffer sync-points
to the non-signalled state, by appending event reset commands to the
end of the user defined command-list. There is also a reset added for
the UR command-buffer wait event that is signalled when the wait list
dependencies of `urCommandBufferEnqueueExp` are met.

Additionally this patch removes the host scope flag from being set on L0 events
created by the UR command-buffer, as we never signal/reset them from host, only
device.