Fix L0 Event leak without return sync point
For out-of-order command-lists in the UR command-buffer
we always create a L0 event to synchronize the commands.

However, this L0 event is only registered with the command-buffer
when there is a return sync-point to associate it with. This means
that if no return sync-point is passed, then the L0 event is
leaked as the command-buffer won't free it on destruction.

Fixed by always registering the L0 event with a sync point, regardless
of if a return sync point is passed.