Avoid throwing when L0 cast fails
When the `dynamic_cast` of a queue object
from a `ur_queue_handle_t` to another type fails, then
an exception is currently thrown.

However, it does not seem to be convention in the adapter code to return errors
by uncaught exceptions. We could instead return an error code, however
I am proposing the modification of this check to an assert as it
seems like this case would signify that the runtime is in an unexpected
state due to programmer error.